Refactor and test.
