+++ e2e-tests/users-custom/conf/backup-minio.yml
@@ -1,0 +1,7 @@
+apiVersion: psmdb.percona.com/v1
+kind: PerconaServerMongoDBBackup
+metadata:
+  name: backup-minio
+spec:
+  clusterName: some-name
+  storageName: minio
