+++ e2e-tests/users-custom/conf/secrets.yml
@@ -1,0 +1,28 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: some-users
+type: Opaque
+data:
+  MONGODB_BACKUP_USER: YmFja3Vw
+  MONGODB_BACKUP_PASSWORD: YmFja3VwMTIzNDU2
+  MONGODB_CLUSTER_ADMIN_USER: Y2x1c3RlckFkbWlu
+  MONGODB_CLUSTER_ADMIN_PASSWORD: Y2x1c3RlckFkbWluMTIzNDU2
+  MONGODB_CLUSTER_MONITOR_USER: Y2x1c3Rlck1vbml0b3I=
+  MONGODB_CLUSTER_MONITOR_PASSWORD: Y2x1c3Rlck1vbml0b3IxMjM0NTY=
+---
+apiVersion: v1
+kind: Secret
+metadata:
+  name: user-one
+type: Opaque
+stringData:
+  userOnePass: 234lkj$oolkjou
+---
+apiVersion: v1
+kind: Secret
+metadata:
+  name: user-two
+type: Opaque
+stringData:
+  userTwoPass: 234lk09#$oolkjou
