+++ e2e-tests/users-custom/run
@@ -1,0 +1,189 @@
+#!/bin/bash
+
+set -o errexit
+
+test_dir=$(realpath $(dirname $0))
+. ${test_dir}/../functions
+set_debug
+
+newpass="test-password"
+newpassencrypted=$(echo -n "$newpass" | base64)
+
+create_infra $namespace
+
+deploy_minio
+
+desc 'create secrets and start client'
+kubectl_bin apply -f "${conf_dir}/client.yml" \
+	-f "${conf_dir}/secrets.yml" \
+	-f "${conf_dir}/minio-secret.yml"
+
+cluster="some-name-rs0"
+desc "create first PSMDB cluster $cluster"
+apply_cluster $test_dir/conf/$cluster.yml
+desc 'Check if all 3 Pods started'
+wait_for_running $cluster 3
+
+backup_name="backup-minio"
+
+desc 'change MONGODB_DATABASE_ADMIN_PASSWORD'
+patch_secret "some-users" "MONGODB_DATABASE_ADMIN_PASSWORD" "$newpassencrypted"
+sleep 25
+
+psmdb="some-name"
+wait_cluster_consistency $psmdb
+sleep 15
+
+user=$(getSecretData "some-users" "MONGODB_DATABASE_ADMIN_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'change MONGODB_BACKUP_PASSWORD'
+patch_secret "some-users" "MONGODB_BACKUP_PASSWORD" "$newpassencrypted"
+sleep 25
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_BACKUP_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'change MONGODB_BACKUP_USER'
+newnameencrypted=$(echo -n "backup2" | base64)
+patch_secret "some-users" "MONGODB_BACKUP_USER" "$newnameencrypted"
+sleep 25
+
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_BACKUP_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'run backup'
+run_backup minio
+wait_backup "$backup_name"
+
+desc 'change MONGODB_USER_ADMIN_PASSWORD'
+patch_secret "some-users" "MONGODB_USER_ADMIN_PASSWORD" "$newpassencrypted"
+sleep 25
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_USER_ADMIN_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'change MONGODB_USER_ADMIN_USER'
+newnameencrypted=$(echo -n "userAdmin2" | base64)
+patch_secret "some-users" "MONGODB_USER_ADMIN_USER" "$newnameencrypted"
+sleep 25
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_USER_ADMIN_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'change MONGODB_CLUSTER_ADMIN_PASSWORD'
+patch_secret "some-users" "MONGODB_CLUSTER_ADMIN_PASSWORD" "$newpassencrypted"
+sleep 25
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_CLUSTER_ADMIN_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'change MONGODB_CLUSTER_MONITOR_PASSWORD'
+patch_secret "some-users" "MONGODB_CLUSTER_MONITOR_PASSWORD" "$newpassencrypted"
+sleep 25
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_CLUSTER_MONITOR_USER")
+check_mongo_auth "$user:$newpass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$newpass@$cluster-2.$cluster.$namespace"
+
+desc 'remove users secret'
+kubectl_bin delete secret some-users
+sleep 35
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_USER_ADMIN_USER")
+pass=$(getSecretData "some-users" "MONGODB_USER_ADMIN_PASSWORD")
+check_mongo_auth "$user:$pass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-2.$cluster.$namespace"
+
+desc 'update all users'
+kubectl_bin apply -f "${conf_dir}/secrets.yml"
+sleep 35
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_USER_ADMIN_USER")
+pass=$(getSecretData "some-users" "MONGODB_USER_ADMIN_PASSWORD")
+check_mongo_auth "$user:$pass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-2.$cluster.$namespace"
+
+desc 'update monitor user username'
+patch_secret "some-users" "MONGODB_CLUSTER_MONITOR_USER" "$(echo -n "$newmonitorusername" | base64)"
+sleep 35
+wait_cluster_consistency $psmdb
+sleep 15
+user=$(getSecretData "some-users" "MONGODB_CLUSTER_MONITOR_USER")
+pass=$(getSecretData "some-users" "MONGODB_CLUSTER_MONITOR_PASSWORD")
+check_mongo_auth "$user:$pass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-2.$cluster.$namespace"
+
+desc 'secret without userAdmin'
+kubectl_bin apply -f "${test_dir}/conf/secrets.yml"
+sleep 25
+user=$(getSecretData "some-users" "MONGODB_USER_ADMIN_USER")
+pass=$(getSecretData "some-users" "MONGODB_USER_ADMIN_PASSWORD")
+check_mongo_auth "$user:$pass@$cluster-0.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-1.$cluster.$namespace"
+check_mongo_auth "$user:$pass@$cluster-2.$cluster.$namespace"
+
+desc 'custom user defined users'
+kubectl_bin apply -f "${test_dir}/conf/secrets.yml"
+
+
+kubectl_bin patch psmdb ${cluster} --type=merge --patch '{
+		"spec": {"users":[
+			{
+				"name":"user-one",
+				"db":"admin",
+				"passwordSecret": {
+					"name": "user-one",
+					"key": "userOnePass"
+				}
+			},
+			{
+				"name":"user-two",
+				"db":"admin",
+				"passwordSecret": {
+					"name": "user-two",
+					"key": "userTwoPass"
+				}
+			},
+		]}
+	}'
+sleep 120
+
+userOne="user-one"
+userOnePass=$(getSecretData "user-one" "userOnePass")
+# userTwoPass=$(getSecretData "user-two" "userTwoPass")
+
+check_mongo_auth "$userOne:$userOnePass@$cluster-0.$cluster.$namespace"
+# check_mongo_auth "$user:$pass@$cluster-1.$cluster.$namespace"
+
+desc 'update custom user defined user passwords'
+
+
+destroy $namespace
+
+desc 'test passed'
